Move button functions on a dedicated file
