+++ src/theme/components/Link.tsx
@@ -59,76 +59,3 @@ export const LinkWithMaterial = MaterialStyled(MaterialLink)<LinkProps>(
   }
 `,
 );
-
-type LinkInButtonProps = Pick<
-  CSSProperties,
-  'backgroundColor' | 'color' | 'borderWidth' | 'padding'
->;
-const ExternalLinkInButton = styled.a<LinkInButtonProps>`
-  padding: 12px 32px;
-  text-align: center;
-  font-size: 16.5px;
-  font-weight: 700;
-  margin: 4px 2px;
-  border-radius: 20px;
-  text-decoration: none;
-  position: relative;
-  background-color: ${props => props.backgroundColor};
-  color: ${props => props.color};
-  border: ${props => `2px solid ${props.backgroundColor}`};
-  display: inline-block;
-
-  &:hover {
-    background-color: ${props => props.color};
-    color: ${props => props.backgroundColor};
-  }
-`;
-
-//padding: 0px 15px;
-export const ButtonWithExternalLink = ({
-  children,
-  ...rest
-}: // eslint-disable-next-line  @typescript-eslint/no-explicit-any
-React.PropsWithChildren<AnchorHTMLAttributes<any>> &
-  LinkInButtonProps): JSX.Element => (
-  <StyledButton>
-    <ExternalLinkInButton {...rest}>{children}</ExternalLinkInButton>
-  </StyledButton>
-);
-
-const StyledButton = styled.div`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-`;
-
-const InternalLinkInButton = styled(GatsbyLink)<LinkInButtonProps>`
-  padding: 12px 32px;
-  text-align: center;
-  font-size: 16.5px;
-  font-weight: 700;
-  margin: 4px 2px;
-  border-radius: 20px;
-  text-decoration: none;
-  position: relative;
-  background-color: ${props => props.backgroundColor};
-  color: ${props => props.color};
-  border: ${props => `2px solid ${props.backgroundColor}`};
-  display: inline-block;
-
-  &:hover {
-    background-color: ${props => props.color};
-    color: ${props => props.backgroundColor};
-  }
-`;
-
-export const ButtonWithInternalLink = ({
-  children,
-  ...rest
-}: React.PropsWithChildren<{ to: string }> &
-  LinkInButtonProps): JSX.Element => (
-  <StyledButton>
-    <InternalLinkInButton {...rest}>{children}</InternalLinkInButton>
-  </StyledButton>
-);
